updated dashboard view

+++ webapp/views/notice/index.vue
@@ -8,6 +8,7 @@ import { onMounted, reactive, computed } from "vue";
 import { INotice } from "../../models/notice";
 import { useNoticeStore, useSetupStore, useAuthStore } from "../../stores";
 import { useApiUtil } from "../../composables";
+import Swal from 'sweetalert2';
 
 let setupStore = useSetupStore();
 const noticeStore = useNoticeStore();
@@ -30,10 +31,25 @@ onMounted(async () => {
   await noticeStore.fetchMyNotices(user.value?.uid!);
 });
 
-function deleteNotice(uid: number): void {
-  withClientMutation(DELETE_NOTICE, { uid }, "deleteNotice").then((payload) =>
-    noticeStore.deleteNotice(payload)
-  );
+async function deleteNotice(uid: number) {
+
+  await Swal.fire({
+    title: 'Are you sure?',
+    text: "You want to delete these notice",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete now!',
+    cancelButtonText: 'No, do not delete!',
+  }).then(async (result) => {
+    if (result.isConfirmed) {
+      withClientMutation(DELETE_NOTICE, { uid }, "deleteNotice").then((payload) =>
+        noticeStore.deleteNotice(payload)
+      );
+    }
+  })
+
 }
 
 function FormManager(create: boolean, obj: INotice = {} as INotice): void {
@@ -51,29 +67,22 @@ const notices = computed<INotice[]>(() => noticeStore.getMyNotices(user.value?.u
 
 <template>
   <h1 class="h1 my-4 font-bold text-dark-700 mr-4">Notice Manager</h1>
-  <button
-    @click.prevent="FormManager(true)"
-    class="px-4 my-2 p-1 text-sm border-sky-800 border text-dark-700 transition-colors duration-150 rounded-sm focus:outline-none hover:bg-sky-800 hover:text-gray-100"
-  >
+  <button @click.prevent="FormManager(true)"
+    class="px-4 my-2 p-1 text-sm border-sky-800 border text-dark-700 transition-colors duration-150 rounded-sm focus:outline-none hover:bg-sky-800 hover:text-gray-100">
     New Notice
   </button>
 
   <!-- Notice Table View -->
   <div class="overflow-x-auto mt-4">
     <div
-      class="align-middle inline-block min-w-full shadow overflow-hidden bg-white shadow-dashboard px-2 pt-1 rounded-bl-lg rounded-br-lg"
-    >
+      class="align-middle inline-block min-w-full shadow overflow-hidden bg-white shadow-dashboard px-2 pt-1 rounded-bl-lg rounded-br-lg">
       <table class="min-w-full">
         <thead>
           <tr>
-            <th
-              class="px-1 py-1 border-b-2 border-gray-300 text-left text-sm leading-4 text-gray-800 tracking-wider"
-            >
+            <th class="px-1 py-1 border-b-2 border-gray-300 text-left text-sm leading-4 text-gray-800 tracking-wider">
               Notice Title
             </th>
-            <th
-              class="px-1 py-1 border-b-2 border-gray-300 text-left text-sm leading-4 text-gray-800 tracking-wider"
-            >
+            <th class="px-1 py-1 border-b-2 border-gray-300 text-left text-sm leading-4 text-gray-800 tracking-wider">
               Expiration
             </th>
             <th class="px-1 py-1 border-b-2 border-gray-300"></th>
@@ -83,10 +92,7 @@ const notices = computed<INotice[]>(() => noticeStore.getMyNotices(user.value?.u
           <tr v-for="notice in notices" :key="notice.uid">
             <td class="px-1 py-1 whitespace-no-wrap border-b border-gray-500">
               <div class="flex items-center">
-                <div
-                  class="text-sm leading-5 text-gray-800"
-                  @click="FormManager(false, notice)"
-                >
+                <div class="text-sm leading-5 text-gray-800" @click="FormManager(false, notice)">
                   {{ notice.title }}
                 </div>
               </div>
@@ -94,19 +100,15 @@ const notices = computed<INotice[]>(() => noticeStore.getMyNotices(user.value?.u
             <td class="px-1 py-1 whitespace-no-wrap border-b border-gray-500">
               {{ notice.status }}
             </td>
-            <td
-              class="px-1 py-1 whitespace-no-wrap text-right border-b border-gray-500 text-sm leading-5"
-            >
+            <td class="px-1 py-1 whitespace-no-wrap text-right border-b border-gray-500 text-sm leading-5">
               <button
                 class="px-2 py-1 mr-2 border-grey-500 border text-grey-500rounded-smtransition duration-300 hover:bg-gray-100 hover:text-black-700 focus:outline-none"
-                @click="FormManager(false, notice)"
-              >
+                @click="FormManager(false, notice)">
                 View/Edit
               </button>
               <button
                 class="px-2 py-1 mr-2 ml-2 border-orange-600 border text-orange-600rounded-smtransition duration-300 hover:bg-orange-600 hover:text-black-700 focus:outline-none"
-                @click="deleteNotice(notice.uid)"
-              >
+                @click="deleteNotice(notice.uid)">
                 Delete
               </button>
             </td>
